// src/types/cardsTypes.ts
// Cards state selectors
export interface cardsStateSelectors {
  liItem: string
  soundDiv: string
  soundImageWrapper: string
  soundImage: string
  soundPlayButton: string
  soundTitle: string
  soundDownloadButton: string
  soundDuration: string
  soundWaveformWrapper: string
  isActive: string
}

// Cards selectors
export interface cardsSelectors {
  soundsList: string
}